Commit 21

+++ src/main/java/gob/jmas/dto/ConceptoDePagoDto.java
@@ -4,11 +4,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
 public class ConceptoDePagoDto {
-    public ConceptoDePagoDto(Integer claveComercial, Integer claveSat, String descripcion, Double monto, Double tasa) {
+    public ConceptoDePagoDto(Integer claveComercial, Integer claveSat, String descripcion, BigDecimal monto, BigDecimal tasa) {
         this.claveComercial = claveComercial;
         this.claveSat = claveSat;
         this.descripcion = descripcion;
@@ -19,6 +21,6 @@ public class ConceptoDePagoDto {
     private Integer claveComercial;
     private Integer claveSat;
     private String descripcion;
-    private Double monto;
-    private Double tasa;
+    private BigDecimal monto;
+    private BigDecimal tasa;
 }
