+++ src/main/java/gob/jmas/model/comercial/Pago.java
@@ -10,6 +10,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -44,7 +45,7 @@ public class Pago {
     private Integer claveComercial;
     private Integer claveSat;
     private String descripcion;
-    private Double monto;
-    private Double tasa;
+    private BigDecimal monto;
+    private BigDecimal tasa;
 
 }
